+++ lib/engineyard-serverside/deploy.rb
@@ -343,23 +343,61 @@ YML
 
       def symlink_configs
         shell.status "Preparing shared resources for release."
+
+        shell.substatus "Set group write permissions"
+        run "chmod -R g+w #{paths.active_release}"
+
+        setup_shared_tmp
+
         symlink_tasks.each do |what, cmd|
           shell.substatus what
           run(cmd)
         end
       end
 
+      def setup_shared_tmp
+        if config.shared_tmp?
+          shell.substatus "Creating and symlinking shared tmp directory if empty"
+          run <<-SETUP_TMP
+if [ -e "#{paths.active_tmp}" ] && [ ! "$(ls #{paths.active_tmp} 2>&1)" ]; then
+  rm -rf #{paths.active_tmp};
+fi;
+if [ ! -e "#{paths.active_tmp}" ]; then
+  mkdir -p #{paths.shared_tmp} && ln -nfs #{paths.shared_tmp} #{paths.active_tmp};
+fi
+          SETUP_TMP
+
+          unless paths.active_tmp.symlink?
+            note = "This application's repository has tmp/ with contents committed.\n"
+            if config.precompile_assets?
+              note << "Asset precompile speed can be increased by sharing tmp across releases.\n"
+            end
+            note << <<-NOTICE
+Enable shared tmp with the follow git command (if applicable):
+
+  $ git rm -r tmp/* && echo '*' > tmp/.gitignore
+
+Or, disable this feature via ey.yml configuration as follows:
+
+defaults:
+  shared_tmp: false
+            NOTICE
+            shell.notice note
+          end
+        else
+          shell.substatus "Prepare shared pids directory"
+          run "rm -rf #{paths.active_tmp}/pids"
+        end
+      end
+
       def symlink_tasks
         [
-          ["Set group write permissions",           "chmod -R g+w #{paths.active_release}"],
+          ["Symlink shared pids directory",         "ln -nfs #{paths.shared}/pids #{paths.active_tmp}/pids"],
+          ["Symlink shared log directory",          "rm -rf #{paths.active_log} && ln -nfs #{paths.shared_log} #{paths.active_log}"],
           ["Remove public/system if symlinked",     "if [ -L \"#{paths.public_system}\" ]; then rm -rf #{paths.public_system}; fi"],
-          ["Remove symlinked shared directories",   "rm -rf #{paths.active_log} #{paths.active_release}/tmp/pids"],
-          ["Create tmp directory",                  "mkdir -p #{paths.active_release}/tmp"],
           ["Create public directory",               "mkdir -p #{paths.public}"],
-          ["Create config directory",               "mkdir -p #{paths.active_release_config}"],
-          ["Symlink shared log directory",          "ln -nfs #{paths.shared_log} #{paths.active_log}"],
           ["Symlink public system directory",       "if [ ! -e \"#{paths.public_system}\" ]; then ln -ns #{paths.shared_system} #{paths.public_system}; fi"],
-          ["Symlink shared pids directory",         "ln -nfs #{paths.shared}/pids #{paths.active_release}/tmp/pids"],
+          ["Create config directory",               "mkdir -p #{paths.active_release_config}"],
           ["Symlink other shared config files",     "find #{paths.shared_config} -maxdepth 1 -type f -not -name 'database.yml' -exec ln -s {} #{paths.active_release_config} \\;"],
           ["Symlink database.yml if needed",        "if [ -f \"#{paths.shared_config}/database.yml\" ]; then ln -nfs #{paths.shared_config}/database.yml #{paths.active_release_config}/database.yml; fi"],
           ["Symlink newrelic.yml if needed",        "if [ -f \"#{paths.shared_config}/newrelic.yml\" ]; then ln -nfs #{paths.shared_config}/newrelic.yml #{paths.active_release_config}/newrelic.yml; fi"],