+++ ChangeLog.md
@@ -5,6 +5,8 @@
   * Add new ey.yml option `precompile_assets_command`
     Setting `precompile_assets_command` overrides the asset precompile rake command. (default: `rake assets:precompile RAILS_GROUPS=assets`)
     Bundler binstubs are in PATH so gem binaries will load through bundler.
+  * In order to speed up asset compilation in Rails, the directory `/data/app/current/tmp` is now preserved between deploys to maintain assets cache.
+    To disable this feature, set `shared_tmp: false` in `ey.yml`.
 
 ## v2.5.0 (2014-09-23)